+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.0-alpha.2] - 2022-01-29
 ### Added
 - DotNet template for toolRestore and restore in one step named restoreWithTools
 - DotNet template for basic packing